Completed the feature 'shared interface type id for identical interfaces'
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@299 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/add_on/scriptfile/scriptfile.cpp
@@ -110,7 +110,7 @@ asCScriptString *asCScriptFile::ReadString(unsigned int length)
 	// Read the string
 	string buf;
 	buf.resize(length);
-	int size = fread(&buf[0], 1, length, file); 
+	int size = (int)fread(&buf[0], 1, length, file); 
 	buf.resize(size);
 
 	// Create the string object that will be returned
